setup basic endpoint for prodcut with stripe check.
I have added some packages and setup basic post endpoint
for the product we check that the request came from stripe,
using private keys setup in the .ENV file that is now setup.
We are then able once we know it was pushed by stripe to perfom actions.
These actions will be setting up the user on cyber panel by making the requests.
This will be added next at the moment on the request nothing will happen.

LainShot :)